+++ packages/server/src/main.ts
@@ -1,8 +1,21 @@
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
+import { INestApplication } from '@nestjs/common';
+import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
 import { constant } from '@hello/common';
+import { AppModule } from './app.module';
 import { HttpExceptionFilter } from './filters/httpException.filter';
 import { TransformResponseInterceptor } from './interceptors/transformResponse.interceptor';
+
+function setApiDocs(app: INestApplication) {
+  const config = new DocumentBuilder()
+    .setTitle('Hello 聊天系统')
+    .setDescription('Hello 聊天系统接口文档')
+    .setVersion('1.0')
+    .addBearerAuth()
+    .build();
+  const document = SwaggerModule.createDocument(app, config);
+  SwaggerModule.setup('api-docs', app, document);
+}
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule, {
@@ -15,6 +28,8 @@ async function bootstrap() {
   app.setGlobalPrefix('api/v1');
   app.useGlobalFilters(new HttpExceptionFilter());
   app.useGlobalInterceptors(new TransformResponseInterceptor());
+
+  setApiDocs(app);
 
   await app.listen(4000);
 }