+++ packages/server/src/modules/user/user.controller.ts
@@ -1,11 +1,14 @@
 import { Body, Controller, Post } from '@nestjs/common';
+import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import { CreateUserDTO } from './dto/createUser.dto';
 import { UserService } from './user.service';
 
+@ApiTags('用户接口')
 @Controller('user')
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
+  @ApiOperation({ summary: '新建用户' })
   @Post()
   async create(@Body() dto: CreateUserDTO) {
     return this.userService.create(dto);