+++ packages/server/config/conf.ts
@@ -1,0 +1,2 @@
+// TOOD: 占时写在这里，应该从环境变量中获取
+export const JWT_SECRET = 'jwt_secret';