+++ packages/server/src/modules/auth/auth.module.ts
@@ -1,0 +1,21 @@
+import { Module } from '@nestjs/common';
+import { PassportModule } from '@nestjs/passport';
+import { JwtModule } from '@nestjs/jwt';
+import { UserModule } from '../user/user.module';
+import { AuthController } from './auth.controller';
+import { AuthService } from './auth.service';
+import { LocalStrategy } from './strategies/local.strategy';
+import { JWT_SECRET } from 'config/conf';
+
+const jwtModule = JwtModule.register({
+  secret: JWT_SECRET,
+  signOptions: { expiresIn: '1h' },
+});
+
+@Module({
+  imports: [UserModule, PassportModule, jwtModule],
+  exports: [],
+  providers: [AuthService, LocalStrategy],
+  controllers: [AuthController],
+})
+export class AuthModule {}