+++ packages/server/src/modules/auth/auth.service.ts
@@ -1,0 +1,31 @@
+import { Injectable } from '@nestjs/common';
+import { JwtService } from '@nestjs/jwt';
+import { UserEntity } from '../user/entities/user.entity';
+import { UserService } from '../user/user.service';
+
+@Injectable()
+export class AuthService {
+  constructor(private readonly userService: UserService, private readonly jwtService: JwtService) {}
+
+  async validateUser(account: string, password: string) {
+    const user = await this.userService.findOne(account);
+
+    if (user && user.password === password) {
+      const { password, ...result } = user;
+
+      return result;
+    }
+
+    return null;
+  }
+
+  generateToken(user: Partial<UserEntity>) {
+    return this.jwtService.sign(user);
+  }
+
+  async login(user: Partial<UserEntity>) {
+    const token = this.generateToken(user);
+
+    return { token };
+  }
+}