+++ packages/server/src/modules/user/user.controller.ts
@@ -1,7 +1,6 @@
-import { Body, Controller, Post, Response } from '@nestjs/common';
+import { Body, Controller, Post } from '@nestjs/common';
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import { CreateUserDTO } from './dto/createUser.dto';
-import { LoginUserDTO } from './dto/loginUser.dto';
 import { UserService } from './user.service';
 
 @ApiTags('用户接口')
@@ -13,11 +12,5 @@ export class UserController {
   @Post('create')
   async create(@Body() dto: CreateUserDTO) {
     return this.userService.create(dto);
-  }
-
-  @ApiOperation({ summary: '登录' })
-  @Post('login')
-  async login(@Body() dto: LoginUserDTO) {
-    return this.userService.login(dto);
   }
 }