+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 
 * @hello/common build ([bcb5083](https://github.com/CCZX/Hello/commit/bcb50837825ea2de245eb86ad800311d9aa7e6c3))
+* axios 封装 ([6e2dd71](https://github.com/CCZX/Hello/commit/6e2dd71f16547b3d9b6a20384db7e66abad4d098))
 * login page ([e78f281](https://github.com/CCZX/Hello/commit/e78f2817de598f6f37eb3dc1ab240e7a89d88a01))
 * user 接口调通 ([0e4ee8f](https://github.com/CCZX/Hello/commit/0e4ee8fdb6981a7bf18f57a03775eb9434158a7e))
 * userEntity ([2f3e45f](https://github.com/CCZX/Hello/commit/2f3e45fec9052b570c6ae2098fc7657fb92ea8dc))